combine event tests and refactor for camera access
